musicdaemon/requirement.txt
python-shout

pytz
python-dateutil

# Database
psycopg2-binary
